+++ src/duckvep/kernel/src/duckvep_effect.h
@@ -61,7 +61,8 @@ typedef enum duckvep_pre_bit {
     DUCKVEP_PRE_CODING_TRANSCRIPT      = 35,
     DUCKVEP_PRE_INFRAME_INSERTION      = 36,
     DUCKVEP_PRE_PROTEIN_ALTERING       = 37,
-    DUCKVEP_PRE_BIT_COUNT              = 38
+    DUCKVEP_PRE_WITHIN_NMD_TRANSCRIPT  = 38,
+    DUCKVEP_PRE_BIT_COUNT              = 39
 } duckvep_pre_bit_t;
 
 #define DUCKVEP_PRE(b) (UINT64_C(1) << (b))